update version and improve settings
