+++ change-notes.html
@@ -1,3 +1,11 @@
+<strong>1.4.13</strong>
+<ul>
+    <li>[FIX]: Fix NPE when building or updating compile commands without a configured toolkit</li>
+    <li>[FIX]: Fix COMPONENTS_LOADED error by separating Project and Application configurables</li>
+    <li>[IMPROVE]: Merge XMake Toolkit settings into the main XMake project settings page</li>
+    <li>[IMPROVE]: Change default setting for auto-update compile_commands.json to false</li>
+    <li>[IMPROVE]: Add meaningful error notifications for build execution failures</li>
+</ul>
 <strong>1.4.12</strong>
 <ul>
     <li>[NEW]: Add C++ Intellisense support by generating compile_commands.json automatically</li>
