+++ gradle.properties
@@ -1,4 +1,4 @@
-pluginVersion=1.4.12
+pluginVersion=1.4.13
 pluginSinceBuild=243
 
 runIdeVersion=2025.3
