+++ src/main/kotlin/io/xmake/project/XMakeSettings.kt
@@ -8,7 +8,7 @@ import com.intellij.openapi.project.Project
 class XMakeSettings : PersistentStateComponent<XMakeSettings.State> {
     data class State(
         var compileCommandsPath: String = "",
-        var autoUpdateCompileCommands: Boolean = true
+        var autoUpdateCompileCommands: Boolean = false
     )
 
     private var myState = State()
